+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ## Changed
 
+- Consolidate `kaocha.hierarchy`, so it can be used for kaocha-cljs
+
 # 0.0-529 (2019-07-04 / 975bbc6)
 
 ## Fixed